% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{com}
\alias{com}
\title{Wrapper for coxph}
\usage{
com(response, time, endpoint, scale = FALSE)
}
\arguments{
\item{response}{The response variable}

\item{time}{Time variable, e.g. time to treatment, time to last known alive...}

\item{endpoint}{Event variable, e.g. treatment, died...}

\item{scale}{Whether to center and scale the response variable}
}
\description{
Perform cox regression using coxph function from survival package
}
