% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{glog}
\alias{glog}
\title{Generalized log transformation}
\usage{
glog(x, q = 0.03)
}
\arguments{
\item{x}{The input matrix or datafram}

\item{q}{The quantile used for transformation}
}
\description{
A function to perform generalized log transformation on a matrix, useful for dataset contians negative values
}
