% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{highCor}
\alias{highCor}
\title{Identify highly correlated variables in a dataset}
\usage{
highCor(X, cut = 0.75, method = "pearson")
}
\arguments{
\item{X}{The input dataframe for matrix. Columns as samples and rows as features.}

\item{cut}{The cutoff of correlation coefficient. The default value is 0.75}

\item{method}{The method used for the calculation of correlation coefficient, either "person" (default) or "spearman".}
}
\description{
A function to identify highly correlated variables.
Return a Boolean vector of highly correlated variables that can be removed from the dataset
}
