% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{km}
\alias{km}
\title{Wrapper for plotting Kaplan-Meier plot}
\usage{
km(response, time, endpoint, titlePlot = "KM plot", pval = NULL,
  stat = "median", maxTime = NULL, showP = TRUE, showTable = FALSE,
  ylab = "Fraction", xlab = "Time (years)", table_ratio = c(0.7,
  0.3))
}
\arguments{
\item{response}{The response variable}

\item{time}{Time variable, e.g. time to treatment, time to last known alive...}

\item{endpoint}{Event variable, e.g. treatment, died...}

\item{titlePlot}{Plot title.}

\item{pval}{The p value that needs to be shown in the plot. If NULL, the p value will be calculated automatically.}

\item{stat}{The method to stratify samples. If the reponse variable is catagorical, only "binary" should be used. If the response variable is continous, either "median" or "maxstat" can be used.}

\item{maxTime}{Truncate the maximal time duration.}

\item{showP}{Whether to show P value in the plot.}

\item{showTable}{Whether to show the "number at risks" table under the KM plot.}

\item{ylab}{Label for y axis.}

\item{xlab}{Label for x axis.}

\item{table_ration}{The ration of height between the KM plot and the table below.}
}
\description{
Using ggsurvplot function from survminer to plot Kaplan-Meier plot
}
