% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{loadData}
\alias{loadData}
\title{Load the newst dataset}
\usage{
loadData(prefix, dir, sep = "_")
}
\arguments{
\item{prefix}{The prefix in the file name}

\item{dir}{The directory of the file}

\item{sep}{The separator the divides prefix and date}
}
\description{
Load the newest data object to R based on the file name
The file name should follow the format as "CPS1000_180126.RData"
}
