% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{loadRData}
\alias{loadRData}
\title{Load multiple data function}
\usage{
loadRData(dataNames, folderDir, verbose = TRUE)
}
\arguments{
\item{dataNames}{A list of the names of the RData files}

\item{folderDir}{The folder that contain the RData files}

\item{verbose}{Whether to show additional information}
}
\description{
This is a function to load a list of data in a specified folder
}
