% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{makepdf}
\alias{makepdf}
\title{Plot a list of ggplot object}
\usage{
makepdf(x, name, ncol = 3, nrow = 2, figNum = NULL, width = 20, height = 12)
}
\arguments{
\item{x}{A list of ggplot object}

\item{name}{The name of the output pdf file}

\item{ncol}{The number of columns in a page}

\item{nrow}{The number of rows in a page}

\item{figNum}{The number of total plots in a page}

\item{width}{The width of the page}

\item{height}{The height of the page}
}
\description{
A function to plot a list of ggplot object in a pdf file
}
