% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mscale}
\alias{mscale}
\title{Calculate modified row z-scores}
\usage{
mscale(x, center = TRUE, scale = TRUE, censor = NULL,
  useMad = FALSE)
}
\arguments{
\item{x}{The input matrix or datafram}

\item{center}{Centered by median}

\item{scale}{Scaled by 1.4826*MAD}

\item{censor}{Whether to censor the scaled value. If a positive integer value, the upper limit will be this value and the lower limit will be the negative value.}

\item{useMad}{Whether use row z score or modified z score (useMad = TRUE)}
}
\description{
Row centered by median and scaled by median absolute deviation
}
