% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{plotHazard}
\alias{plotHazard}
\title{Forest plot for multivariate Cox regression result.
Plot the hazard ration and confidence interval for each risk factors from a multivariate Cox regression results.}
\usage{
plotHazard(survRes, title = "")
}
\arguments{
\item{survRes}{The output of runCox function}

\item{title}{The plot for the title.}
}
\description{
Forest plot for multivariate Cox regression result.
Plot the hazard ration and confidence interval for each risk factors from a multivariate Cox regression results.
}
