% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEP.R
\name{runCamera}
\alias{runCamera}
\title{Function to run enrichment analysis using Camera from LIMMA}
\usage{
runCamera(exprMat, design, gmtFile, id = NULL, contrast = ncol(design),
  method = "camera", pCut = 0.05, ifFDR = FALSE,
  removePrefix = NULL, plotTitle = "", insideLegend = FALSE)
}
\arguments{
\item{exprMat}{Gene expression matrix. Genes as rows and samples as columns.}

\item{design}{Design matrix as used by limma.}

\item{gmtFile}{A path to the gene signature file.}

\item{id}{A character vector of identifiers for genes used in the gmt file, ff the row names are not the same as the identfiers in the gmt file.}

\item{contrast}{The index of the colum in design matrix that should be used as contrast for enrichment analysis. Default is the last column.}

\item{method}{Wheter using "camera" (competitive) or "fry" (self-contained) method}

\item{pCut}{Cut-off for significance of enrichment.}

\item{removePrefix}{Remove a certain string in the set names to make them shorter. For example, "HALLMARK_".}

\item{plotTitle}{The tilte used for enrichment bar plot.}

\item{ifDFR}{Whether raw p value or adjusted p value (FDR) should be used for significance cut-off}

\item{insideLegned}{Whether to put the legend inside the bar plot to save space.}
}
\description{
A funtion to perform gene enrichment analysis using rotational test (fry or camera) implemented in limma package.
}
