% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{runCox}
\alias{runCox}
\title{Multivariate Cox regression}
\usage{
runCox(survTab, riskTab, time, endpoint, id = "patientID")
}
\arguments{
\item{survTab}{A tidy table contain the survival data in three columns: identifier, time, event}

\item{riskTab}{A tidy table that contain the risk factors that need to be tested. One column must contain the sample identifiers. All other columns will be treated as risk factors and tested in the multivariate Cox model.}

\item{time}{The column name of the time variable in the survTab.}

\item{endpoint}{The column name of the event variable in the survTab.}

\item{id}{The column name of the sample identifier, which is used to join the two table.}
}
\description{
Wrapper function to run multivariate Cox regression using coxph
}
