% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sumToTiday}
\alias{sumToTiday}
\title{Convert SummarizedExperiment object to a tidy table}
\usage{
sumToTiday(seObject, rowID = "rowID", colID = "colID")
}
\arguments{
\item{seObject}{Input SummarizedExperiment object.}

\item{rowID}{A character variable specifying the name of row identifier.}

\item{colID}{A character variable specifying the name of column identifier.}
}
\description{
Function for converting a SummarizedExperiment object to a tidy table.
}
