% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tidyToSum}
\alias{tidyToSum}
\title{Convert tidy table to SummarizedExperiment object}
\usage{
tidyToSum(tidyTable, rowID, colID, values, annoRow, annoCol)
}
\arguments{
\item{tidyTable}{Input tidy table}

\item{rowID}{A character variable specifying the column in the tidy table that should be used as row identifiers in the SE object.}

\item{colID}{A character variable specifying the column in the tidy table that should be used as column identifiers in the SE object.}

\item{values}{A character or a vector of characters, specifying the columns that should be treated as measurements (assays).}

\item{annoRow}{A character or a vector of characters, specifying the columns that should be used a rowData}

\item{annoCol}{A character or a vector of characters, specifying the columns that should be used a colData.}
}
\description{
Function for converting a tidy table to SummarizedExperiment object.
}
